+++ src/supla/sensor/general_purpose_channel_base.cpp
@@ -309,7 +309,7 @@ void GeneralPurposeChannelBase::setChannelRefreshIntervalMs(
                   refreshIntervalMs);
 }
 
-void GeneralPurposeChannelBase::setRefreshIntervalMs(int intervalMs,
+void GeneralPurposeChannelBase::setRefreshIntervalMs(int32_t intervalMs,
                                                      bool local) {
   // RefreshIntervalMs in config contain value from server/Cloud. We don't
   // adjust it to internal value.
@@ -321,7 +321,7 @@ void GeneralPurposeChannelBase::setRefreshIntervalMs(int intervalMs,
   auto oldIntervalMs = getRefreshIntervalMs();
   commonConfig.refreshIntervalMs = intervalMs;
   setChannelRefreshIntervalMs(intervalMs);
-  if (intervalMs != oldIntervalMs && local) {
+  if (static_cast<uint16_t>(intervalMs) != oldIntervalMs && local) {
     channelConfigState = Supla::ChannelConfigState::LocalChangePending;
     saveConfig();
     saveConfigChangeFlag();
